refactor: domain - Delete Immutablelines and more
