+++ src/main/java/subway/view/screen/info/ViewSubway.java
@@ -1,7 +1,6 @@
 package subway.view.screen.info;
 
 import java.util.List;
-import subway.domain.ImmutableLine;
 import subway.domain.Line;
 import subway.domain.Subway;
 import subway.view.message.InfoMessage;
@@ -11,9 +10,9 @@ public class ViewSubway implements ViewUI {
 
     @Override
     public void commands(Subway subway) {
-        List<ImmutableLine> lineList = subway.getLineList();
+        List<Line> lineList = subway.getLineList();
 
-        for (ImmutableLine line : lineList) {
+        for (Line line : lineList) {
             System.out.println(InfoMessage.getInfoHeader() + line.getName());
             System.out.println(InfoMessage.getInfoHeader() + "---");
             OutputTool.printStringListWithHeader(line.getLineStationNameList(),
